Moved neon/http interfaces to new file
